world_hello/requirements.txt
google-adk>=1.21.0
python-dotenv>=1.0.0
google-cloud-aiplatform>=1.80.0

